src/main/webapp/app/entities/taisannhadatid/delete/taisannhadatid-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { TaisannhadatidService } from '../service/taisannhadatid.service';

import { TaisannhadatidDeleteDialogComponent } from './taisannhadatid-delete-dialog.component';

describe('Taisannhadatid Management Delete Component', () => {
  let comp: TaisannhadatidDeleteDialogComponent;
  let fixture: ComponentFixture<TaisannhadatidDeleteDialogComponent>;
  let service: TaisannhadatidService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [TaisannhadatidDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(TaisannhadatidDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(TaisannhadatidDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(TaisannhadatidService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
